+++ src/datasources/results/components/query-builders/query-results/ResultsQueryBuilder.tsx
@@ -172,6 +172,8 @@ export const ResultsQueryBuilder: React.FC<ResultsQueryBuilderProps> = ({
       QueryBuilderOperations.LIST_DOES_NOT_EQUAL,
       QueryBuilderOperations.LIST_CONTAINS,
       QueryBuilderOperations.LIST_DOES_NOT_CONTAIN,
+      QueryBuilderOperations.DATE_TIME_IS_AFTER,
+      QueryBuilderOperations.DATE_TIME_IS_BEFORE,
     ].map(operation => {
       return {
         ...operation,